release v0.5.0: MMFF performance, parameter accuracy, and benchmark/audit tooling
Performance:
- Cache all term parameters in MMFFForceField::new (E+G 6.2x faster;
  8.6x -> 1.4x vs RDKit; faster than RDKit on simpler sets)
- Build: opt-level=3 + wasm32-simd128 (.cargo/config.toml); ~1.28x WASM speedup

Parameter accuracy:
- Add P_4-C_AR / P_3-C_AR bond params (triphenylphosphine_oxide +9.4 -> 0.00)
- Add CR3R-P_3 bond param (phosphirane; full-precision RDKit audit)
- Remove 11 dead duplicate bond match arms + clippy lints (0 warnings)
- CB_OVERRIDES: cb=1.0 is universal MMFF cs=-2.0 (verified empirically)

Tooling:
- benchmark_mmff.py: 230-mol accuracy+speed gate (exit 1 on regression)
- bench_mmff.rs, dump_params.rs, audit_params.py
- val_set_new6: regression gate for aryl phosphines

Validation: 230/230 match RDKit <0.01 kcal/mol; cargo test 187 pass; clippy 0; WASM builds.

+++ src/mmff/charges.rs
@@ -1064,9 +1064,7 @@ fn compute_mmff_formal_charges(mol: &Molecule, type_ids: &[u8]) -> Vec<f64> {
                     }
                 }
                 // Sum formal charges of all conjugated N atoms and divide
-                let total_fc: f64 = conj.iter()
-                    .map(|&ci| mol.atoms[ci].charge as f64)
-                    .sum();
+                let total_fc: f64 = conj.iter().map(|&ci| mol.atoms[ci].charge).sum();
                 total_fc / conj.len() as f64
             }
             // Halide anions
